# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: YKS Soru Takibi API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tests/{id}/progress:
    get:
      operationId: getTestProgress
      tags: [tests]
      summary: Get test session progress
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Test progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestSessionProgress"
    put:
      operationId: saveTestProgress
      tags: [tests]
      summary: Save or update test session progress
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveTestProgressInput"
      responses:
        "200":
          description: Progress saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestSessionProgress"
  /resources:
    get:
      operationId: listResources
      tags: [resources]
      summary: List all resources
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: lesson
          in: query
          schema:
            type: string
        - name: resourceType
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of resources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResourceWithStats"
    post:
      operationId: createResource
      tags: [resources]
      summary: Create a new resource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResourceInput"
      responses:
        "201":
          description: Resource created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
  /resources/{id}:
    get:
      operationId: getResource
      tags: [resources]
      summary: Get resource details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resource details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceWithStats"
    put:
      operationId: updateResource
      tags: [resources]
      summary: Update a resource
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateResourceInput"
      responses:
        "200":
          description: Resource updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
    delete:
      operationId: deleteResource
      tags: [resources]
      summary: Delete a resource
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Resource deleted
...
components:
  schemas:
...
    TestSessionProgress:
      type: object
      properties:
        id:
          type: integer
        testSessionId:
          type: integer
        currentIndex:
          type: integer
        timer:
          type: integer
          nullable: true
        elapsed:
          type: integer
        isCompleted:
          type: boolean
        completedAt:
          type: string
          nullable: true
        inlineDrawEnabled:
          type: boolean
        collapsedLessons:
          type: object
          additionalProperties:
            type: boolean
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - testSessionId
        - currentIndex
        - elapsed
        - isCompleted
        - inlineDrawEnabled
        - createdAt
        - updatedAt
    SaveTestProgressInput:
      type: object
      properties:
        currentIndex:
          type: integer
        timer:
          type: integer
          nullable: true
        elapsed:
          type: integer
        isCompleted:
          type: boolean
        inlineDrawEnabled:
          type: boolean
        collapsedLessons:
          type: object
          additionalProperties:
            type: boolean
          nullable: true
    Resource:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        publisher:
          type: string
          nullable: true
        category:
          type: string
        lesson:
          type: string
          nullable: true
        topic:
          type: string
          nullable: true
        resourceType:
          type: string
        targetQuestionCount:
          type: integer
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - category
        - resourceType
        - createdAt
        - updatedAt
    ResourceWithStats:
      allOf:
        - $ref: "#/components/schemas/Resource"
        - type: object
          properties:
            totalQuestions:
              type: integer
            solvedQuestions:
              type: integer
            correctQuestions:
              type: integer
            wrongQuestions:
              type: integer
            totalNotes:
              type: integer
            topicStats:
              type: array
              items:
                type: object
                properties:
                  topic:
                    type: string
                  lesson:
                    type: string
                  total:
                    type: integer
                  solved:
                    type: integer
                  correct:
                    type: integer
                  wrong:
                    type: integer
                required:
                  - topic
                  - lesson
                  - total
                  - solved
                  - correct
                  - wrong
    CreateResourceInput:
      type: object
      properties:
        name:
          type: string
        publisher:
          type: string
          nullable: true
        category:
          type: string
        lesson:
          type: string
          nullable: true
        topic:
          type: string
          nullable: true
        resourceType:
          type: string
        targetQuestionCount:
          type: integer
          nullable: true
      required:
        - name
        - category
        - resourceType
    UpdateResourceInput:
      type: object
      properties:
        name:
          type: string
        publisher:
          type: string
          nullable: true
        category:
          type: string
        lesson:
          type: string
          nullable: true
        topic:
          type: string
          nullable: true
        resourceType:
          type: string
        targetQuestionCount:
          type: integer
          nullable: true