# lib/api-zod/src/generated/types/createResourceInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * YKS Soru Takibi API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateResourceInput {
  name: string;
  publisher?: string | null;
  category: string;
  lesson?: string | null;
  topic?: string | null;
  resourceType: string;
  targetQuestionCount?: number | null;
}