# lib/api-zod/src/generated/types/resource.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * YKS Soru Takibi API
 * OpenAPI spec version: 0.1.0
 */

export interface Resource {
  id: number;
  name: string;
  publisher?: string | null;
  category: string;
  lesson?: string | null;
  topic?: string | null;
  resourceType: string;
  targetQuestionCount?: number | null;
  createdAt: string;
  updatedAt: string;
}