# lib/api-zod/src/generated/types/resourceWithStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * YKS Soru Takibi API
 * OpenAPI spec version: 0.1.0
 */
import type { Resource } from "./resource";
import type { ResourceWithStatsTopicStatsItem } from "./resourceWithStatsTopicStatsItem";

export type ResourceWithStats = Resource & {
  totalQuestions?: number;
  solvedQuestions?: number;
  correctQuestions?: number;
  wrongQuestions?: number;
  totalNotes?: number;
  topicStats?: ResourceWithStatsTopicStatsItem[];
};